feat: add Nitro event handler
