+++ package.json
@@ -62,6 +62,11 @@
     "vitest": "^2.1.1",
     "vue-tsc": "^2.1.6"
   },
+  "build": {
+    "externals": [
+      "defu"
+    ]
+  },
   "stackblitz": {
     "installDependencies": false,
     "startCommand": "pnpm install && pnpm dev:prepare && pnpm dev"
