+++ src/types.ts
@@ -38,4 +38,12 @@ export interface HttpClientHintsOptions {
    * @see https://wicg.github.io/responsive-image-client-hints
    */
   critical?: CriticalClientHintsConfiguration
+  /**
+   * Enable server images (expose Nitro event handler)?.
+   *
+   * The Nitro event handler will export the `httpClientHints` object in the event context.
+   *
+   * If set to `true`, the event handler will apply to `/\.(png|jpeg|jpg|webp|avi)$/`.
+   */
+  serverImages?: true | RegExp | RegExp[]
 }
